trial view-only permissions must reject proband list operations
manipulating the proband list is an example of operations that work with
records of different phoenix modules (trial and proband module respectively)

users with a trial view-only permission level can edit a trial subject list,
if they eg. have the proband permission level to edit any proband.

while this is disjunktive-per-module appraoch the general intention
with permissions levels, it can be undesired. this change shows how to restrict
trial view-only permissions absolutely.

the same affects also operations to edit eCRF forms or changing an eCRF status
to eg. sign it. But for those situations, a team member record with the proper
flag (eg. "sign") is required for the person of the user.

The same flag is about to be introduced for proband list operations, so this
change wont't be needed.