trial view-only permissions must reject proband-realted operations
manipulating the proband list is an example of operations that work with
records of different phoenix modules (trial and proband module respectively).

users with a trial view-only permission level can edit a trial subject list,
if they eg. have the proband permission level to edit any proband. while trying
stay disjunctive per module, ignoring the checks for view-only permissions is
in fact wrong, since "view-only" must not render the same as an "edit-any"
trial permission.

the same affects also operations to edit eCRF forms or changing an eCRF status
to eg. sign it. But for those situations it is less significant, as anyway a
team member record with the proper flag (eg. "sign") is required for the person
of the user. (A similar flag is about to be introduced for proband list soon).

the team member operations show a similar sitation with permissions of trial vs.
permissions of person/organisation module, but this was already considered
properly.

all relevant permissions of the trial service have been revised. for the other
services we keep it as is for now.